@@ -70,12 +70,15 @@ namespace Tauron
         public static IDisposable ToActor<TMessage>(this IObservable<TMessage> source, Func<IActorRef> target)
             => source.SubscribeWithStatus(m => target().Tell(m));
 
-        public static IDisposable ToActor<TMessage>(this IObservable<TMessage> source,
-            Func<IActorContext, IActorRef> target)
+        public static IDisposable ToActor<TMessage>(this IObservable<TMessage> source, Func<IActorContext, IActorRef> target)
             => source.SubscribeWithStatus(m => target(ObservableActor.ExposedContext).Tell(m));
 
         public static IDisposable ToActor<TMessage>(this IObservable<TMessage> source, Func<TMessage, IActorRef> target)
             => source.SubscribeWithStatus(m => target(m).Tell(m));
+
+        public static IDisposable ToActor<TMessage>(this IObservable<TMessage> source, Func<TMessage, IActorRef> target, Func<TMessage, object> selector)
+            => source.SubscribeWithStatus(m => target(m).Tell(selector(m)));
+
 
 
         public static IDisposable ForwardToParent<TMessage>(this IObservable<TMessage> source)
@@ -96,13 +99,14 @@ namespace Tauron
         public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source, Func<IActorRef> target)
             => source.SubscribeWithStatus(m => target().Forward(m));
 
-        public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source,
-            Func<IActorContext, IActorRef> target)
+        public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source,            Func<IActorContext, IActorRef> target)
             => source.SubscribeWithStatus(m => target(ObservableActor.ExposedContext).Forward(m));
 
-        public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source,
-            Func<TMessage, IActorRef> target)
+        public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source, Func<TMessage, IActorRef> target)
             => source.SubscribeWithStatus(m => target(m).Forward(m));
+
+        public static IDisposable ForwardToActor<TMessage>(this IObservable<TMessage> source, Func<TMessage, IActorRef> target, Func<TMessage, object> selector)
+            => source.SubscribeWithStatus(m => target(m).Forward(selector(m)));
 
         #endregion
 
@@ -122,6 +126,9 @@ namespace Tauron
 
         public static IDisposable SubscribeWithStatus<TMessage>(this IObservable<TMessage> source, Action<TMessage> onNext)
             => SubscribeWithStatus(source, null, onNext);
+
+        public static IDisposable SubscribeWithStatus<TMessage>(this IObservable<TMessage> source)
+            => SubscribeWithStatus(source, null, _ => {});
 
         #endregion
     }
