@@ -1,7 +1,10 @@
-﻿using Autofac;
+﻿using System;
+using Autofac;
+using JetBrains.Annotations;
 
 namespace Tauron.Application.Settings
 {
+    [PublicAPI]
     public sealed class SettingsConfiguration
     {
         private readonly ContainerBuilder _builder;
@@ -12,6 +15,13 @@ namespace Tauron.Application.Settings
             where TType : ISettingProviderConfiguration
         {
             _builder.RegisterType<TType>().As<ISettingProviderConfiguration>();
+            return this;
+        }
+
+        public SettingsConfiguration WithProvider<TType>(Func<TType> setting)
+            where TType : ISettingProviderConfiguration
+        {
+            _builder.Register(_ => setting()).As<ISettingProviderConfiguration>();
             return this;
         }
     }
