@@ -14,17 +14,10 @@ namespace Akka.MGIHelper
             var builder = ActorApplication.Create(args);
 
             builder
-                //.ConfigureAkka(context =>
-                //               {
-                //                   //var config = new AkkaRootConfiguration();
-                //                   //config.Akka.ElementAcessor.GetAddElement<DispatcherConfiguration>("CallingThread").Type = typeof(CallingThreadDispatcherInternalConfigurator);
-
-                //                   //return config.CreateConfig();
-                //               })
-                .ConfigureLogging((context, configuration) => configuration.ConfigDefaultLogging("MGI_Helper").WriteTo.Sink<SeriLogViewerSink>())
+                .ConfigureLogging((_, configuration) => configuration.ConfigDefaultLogging("MGI_Helper").WriteTo.Sink<SeriLogViewerSink>())
                 .ConfigureAutoFac(cb => cb.RegisterModule<MainModule>())
-                .ConfigurateAkkaSystem((context, system) => system.RegisterLocalization())
-                .UseWpf<MainWindow>(c => c.WithAppFactory(() => new App()));
+                .ConfigurateAkkaSystem((_, system) => system.RegisterLocalization())
+               .UseWpf<MainWindow, App>();
 
             using var app = builder.Build();
             await app.Run();
