@@ -1,5 +1,5 @@
 ﻿using System.Diagnostics;
-using Tauron.Application.Wpf;
+using Tauron.Application.CommonUI;
 
 namespace Akka.MGIHelper.UI.FanControl
 {
